fix bug long gregorian date

+++ js/date.js
@@ -76,7 +76,7 @@ var dateTime = {
         if (lang === 'ar') {
             return addZero(this.getCurrentDay()) + "/" + this.getCurrentMonth() + "/" + date.getFullYear();
         }
-        var options = {weekday: "long", year: "numeric", month: "long", day: "numeric"}
+        var options = {weekday: "long", year: "numeric", month: "short", day: "numeric"}
         try {
             return date.toLocaleString(lang, options).firstCapitalize();
         } catch (e) {
